Don't subtract pages_shared from amount saved
According to the docs, pages_sharing already takes into account the
pages that are being shared:

    pages_sharing: how many more sites are sharing them i.e. how
    much saved.

Note in particular that it states, "how many *more* sites". In other
words, that number should already have subtracted out the pages that
are being shared, and includes only the additional sites that are
sharing them, so would indeed be the amount saved -- just as it says.

This fixes an issue where the amount saved is under-reported and could
sometimes even be negative.